requirements.txt
crewai
langchain_google_genai
load_dotenv
feedparser